moved toRetryPolicy to RetryOptionsUtils
